core: Add base class for template import
This patch adds ImportVmTemplateCommandBase class
and it has most of the functionality that used
to be in ImportVmTemplateCommand.

This class has 2 derived classes:
- ImportVmTemplateCommand
- ImportVmTemplateFromOvaCommand

This split is important, because some code
for importing template from data/export domain
does not make sense when importing from OVA.

As a side effect, it is now possible to import
template from OVA, because the invalid code is
not executed.

Change-Id: If2c766cb5c9689d3ce9a900512ff9aa9682114d0
Bug-Url: https://bugzilla.redhat.com/1789737
Signed-off-by: Andrej Krejcir <akrejcir@redhat.com>