restapi: Remove Support for API Version 3
Version 3 of the API has been deprecated for long
enough to now be completely removed. This patch
removes support for version 3 requests:

  (GET/POST/PUT/DELETE .../api/v3/...)

And the use of v3-structured types as input or output
to API requests.

For backwards compatibility, Version 4.X (the current
version) of the oVirt REST-API will remain usable by
both:

   GET .../api/vms

and

   GET .../api/v4/vms

This will be the only remanent of the splitting between
version 3 and version 4

Bug-Url: https://bugzilla.redhat.com/1753889
Change-Id: Ie53bab61efeff92adb8fbc51d399541b5d00cf59
Signed-off-by: Ori_Liel <oliel@redhat.com>