Search: fix support of underscore in value
This bugs is a regression. Code was working with PG8 and PG9 but when we
moved to PG10 Underscores stopped to function.

This patch removes all PG8 / PG9 deprecated behaviour and as a result
handles correctly underscores in the search value which should have
backslash before in order to be considered as a character(because an
underscore in PG means "any character" in string matching)

This should be checked also with PG12 to insure that there are no
changes that might cause this BZ to be reproduced again.

Change-Id: Ic9bb89731c18e1af3f7efc81236fbe7cd98afd3b
Bug-Url: https://bugzilla.redhat.com/1710724
Signed-off-by: emesika <emesika@redhat.com>