core: AttachmentDao - decouple dao retrievals from mapper
Encapsulate call to qos dao within a RowMapper. This change wil
facilitate using an alternate implementation for the qos info retrieval
in later patches.

In order to utilize a single NetworkAttachmentRowMapper object and avoid
instantiating mappers in each call the call to
dnsResolverConfigurationDao has also been encapsulated within a
RowMapper which guarantees that when the lambda is called the dns dao
has been injected.

Without the lambda dnsResolverConfigurationDao would have had to be
injected into the field reference of NetworkAttachmentRowMapper upon its
declaration (which is were it is instantiated too). But at this time the
dnsResolverConfigurationDao might still be null because the order of
injection is not specified by the javax.inject library, and so the
attachment mapper might have been instantiated without an actual
dnsResolverConfigurationDao object reference.

Verify that in the dao test for #getAllHosts there are no qos entities
related to the retrieved attachments.

Change-Id: Iaf94cb8f22abf54bc2d909505df99a213bfaf744
Bug-Url: https://bugzilla.redhat.com/1811865
Signed-off-by: Eitan Raviv <eraviv@redhat.com>