ansible: Don't start libvirtd.service
4.4 hosts use socket activation to start libvirt services. In order
to be able to migrate VMs, libvirtd-tls.socket service must be
started. However, this service can't be started if libvirtd.service
is already running. We mustn't start libvirtd.service for this
reason.

Actually, starting libvirtd.service is not necessary if vdsmd.service
is started after it, since supervdsmd service (and thus vdsmd) depends
on it on 4.3 and 4.2 hosts. It will be started automatically when
vdsmd or supervdsmd services are started.  And socket activation is
used on 4.4 hosts.

Change-Id: I3357156d0ed3de73464d35500f498227b92fe309
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>