engine: duplicate disk entries after vm import
After importing a VM with at least one disk and
with one snapshot at least from an export domain,
to the same storage domain, the Disks tab in the
VM view shows two disk entries instead of one
in case VM with one disk.

The root cause of this bug is when reading
the ovf file of the exported VM, the Reader
sets all disk images as Active images
including snapshots.
(snapshots are not Active images)

Now disk images will be Active if
SnapshotType.ACTIVE or SnapshotType.PREVIEW,
otherwise, Active will be set to false.

Change-Id: I4b40e9d81700a3685a311ce63493facd6168543d
Bug-Url: https://bugzilla.redhat.com/1734409
Signed-off-by: Ahmad Khiet <akhiet@redhat.com>