Reconfigure Gluster in Reinstall Model
This patch helps users to reconfigure gluster nodes
by enabling a corrupted node to add it back to gluster peer.
This is achieved by a ansible-playbook.
A condition of gluster being supported in the cluster
and a minimum number of 3 host in the same gluster supported cluster
needs to be satisfied to run the playbook.

Bug-url: https://bugzilla.redhat.com/show_bug.cgi?id=1633126
Change-Id: I7d8f9994d3354807a9dc2fb2f723da1d1982a266
Signed-off-by: Prajith Kesava Prasad <pkesavap@redhat.com>