core: move init status to running vm
This change will move the is_initialization
flag change in vm_static from when the VM
is starting to run to the end of the running process.

We are uncertain if the VM did initialize successfully
when we just starting running a VM. Therefore, with this
commit, now only when the VM changed to UP status, the
flag will be changed in the DB.

This is useful since we have the rerun mechanism, if
a VM is failing to start we would want it to try again
with the same initialization we had in the first time.

Change-Id: I0f36792ee89bb0cbf2ea6c4fcb8fbccf41a11ff2
Bug-Url: https://bugzilla.redhat.com/1780022
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>