core: InterfaceDao - host: less QOS dao calls
When the result set for #getAllInterfacesForVds is mapped to objects, a
call is made via the qos dao to the db to retrieve the corresponding qos
entity. This loads the db with calls in the order of
O(count(interface)).

In this patch a stored procedure which embeds the qos info inside the
interface result set is employed in #getAllInterfacesForVds.
A corresponding mapper is introduced to extract the qos info from the
result set and create the required host network qos entity.

This change reduces the number of calls to the qos dao to zero.

Notes:
* The mapper checks for the existence of a qos id in the result se
  since some interfaces may not have a qos related to them.
* The mapper sets the id of the qos entity to the id of the interface in
   accordance with the behavior of HostNetworkQosDao.

Change-Id: I330617db99a311e183eb812a052bd052f97287e3
Bug-Url: https://bugzilla.redhat.com/1811865
Signed-off-by: Eitan Raviv <eraviv@redhat.com>