core: InterfaceDao - cluster: less QOS dao calls
When the result set for #getAllInterfacesByClusterId is mapped to
objects, a call is made via the qos dao to the db to retrieve the
corresponding qos entity for each interface in the result set. This
loads the db with extra calls in the order of O(count(interface)).

In this patch a stored procedure which embeds the qos info inside the
interface result set is employed in #getAllInterfacesByClusterId. A
corresponding mapper is employed to extract the qos info from the result
set and create the required host network qos entity.

This change reduces the number of calls from getAllInterfacesByClusterId
to the qos dao to zero.

Notes:
* The mapper checks for the existence of a qos id in the result set
  since some interfaces may not have a qos related to them.
* The mapper sets the id of the qos entity to the id of the interface in
  accordance with the behavior of HostNetworkQosDao.

Change-Id: Ie564365fd1c999f75a199810b1523450c5e49cd8
Bug-Url: https://bugzilla.redhat.com/1811866
Signed-off-by: Eitan Raviv <eraviv@redhat.com>