core: AttachmentDao - get by nic+net: less QoS dao calls
When the result set for #getNetworkAttachmentByNicIdAndNetworkId is
mapped to objects, a call is made via the qos dao to the db to retrieve
the corresponding qos entity for each interface in the result set. This
loads the db with an extra call.

In this patch #getNetworkAttachmentByNicIdAndNetworkId uses a store
procedure which embeds the qos info inside the network attachment result
set. A corresponding mapper is employed to extract the qos info from the
result set and create the required host network qos entity.

This change eliminates the call to the qos dao.

Notes:
* The mapper checks for the existence of a qos id in the result set
  since some attachments may not have a qos related to them.
* The mapper sets the id of the qos entity to the id of the attachment
  in accordance with the behavior of HostNetworkQosDao.

Change-Id: Ib0e4b1cc51f5dabfd46ae79616066bdaed47d6eb
Bug-Url: https://bugzilla.redhat.com/1811869
Signed-off-by: Eitan Raviv <eraviv@redhat.com>