core: add indication whether a backup disk participates in a VM backup
Add 'checkpoint' = true/false for each disk that participates
in a VM backup.
This information will guide VDSM on which disk should be included in the
checkpoint creation for incremental VM backup.

The disk will only be included in the checkpoint if the disk format is
'cow', since we cannot add bitmaps to 'raw' disks.

So now the request data from Engine to VDSM contains the following
structure:

{
    "backup_id": "backup_1_id",
    "disks": [
        {
            "img_id": "disk1",
            "domain_id": "domain1",
            "volume_id": "volume1",
            "checkpoint": true/false
        },
        {
            "img_id": "disk2",
            "domain_id": "domain2",
            "volume_id": "volume2",
            "checkpoint": true/false
        },
    "from_checkpoint_id": "from_checkpoint_id",
    "to_checkpoint_id": "to_checkpoint_id"
}

Change-Id: I725bf215ecac0a64d9db3d8b55585b50bbe6a37e
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>