ansible: Add warning about OVN on unsuccessful ansible
If the ovn-controller on the host does not reregister itself,
there might be a phantom chassis entry in the OVN Southbound
DB, which might cause broken OVN flows.
For this reason the user should get a hint, that he might want
to clean up manually when the host is unreachable when the user
tries to remove it.

Change-Id: Ia33f0f35ea80d3fb30b6e2700b9f70060c548a7e
Bug-Url: https://bugzilla.redhat.com/1758289
Signed-off-by: Dominik Holler <dholler@redhat.com>