search: fixing regression in cluster search
This patch fixes a regression caused by [1] that adds some cluster
fields for sorting purpose. Among other fields a "datacenter" field was
added. This cause ambigutiy it the search engine code since "Datacenter"
is also a reference object that can be added in order to search by its
properties.

This patch removes the "datacenter" field from the cluster search fields
leaving "Datacenter" to be used as a regerence object as before so now
we can write in the Cluster TAB "Datacenter." and get all its legal
fields in the open pop-up

[1] https://gerrit.ovirt.org/#/c/100853/

Change-Id: I7dbfcb1a903bbbe61f4cc815913bc64e8840f490
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1789262
Signed-off-by: emesika <emesika@redhat.com>