webadmin: Memory Parsing Failure in French
When creating a Virtual Machine where the language
is French, the size format is set to Mo (Mega Octects).
When modifying the memory size and then pressing tab
as an example, the value is reset to 0. This is because
the parsing expects the text to contain MB as per most
other languages whereas in French it uses Mo as the suffix.

The fix was to modify the parsing so that the proper suffix
comparison per language would be used such as for French, Mo
would be fetched from the message as a configuration so that
other languages that use a different suffix may also benefit
from the change.

Also added support for MiB and localized all combinations in
order to remove the hard coding of the suffixes.

Change-Id: I16779ea1a0b18d8796bfe597ced9a3500d2f5bf3
Bug-Url: https://bugzilla.redhat.com/1719990
Signed-off-by: Steven Rosenberg <srosenbe@redhat.com>