engine: fix 'no updates found' message when there are available updates
This patch fixes 'no updates' found message for RHEL8 and RHVH hosts,
taking in consideration the new yum parsing of the taskNode() method,
and the fact that yum list: updates doesn't work for obsolete.
The module shell: yum check-update is being used now, and its result
is saved in a file which is deleted afterwards.
In additional, check-mode is no longer used and therefore isn't checked for anymore

This fix should be replaced by using the right yum ansible module properly once it
is written and merged to the ansible project.

Change-Id: If5e61a358ca81af77000ff1b60a8f7c26a0097b1
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1779588
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1795184
Signed-off-by: Dana Elfassy <delfassy@redhat.com>