webadmin: fix layout on GWT app loading skeleton
The initial GWT loading page skeleton has been updated:
  - The white line between the masthead and left nav / content area
    has been removed
  - The loading spinner is centered in the content area
  - The scroll bar in the content area holding the loading spinner
    has been removed

Created `_variables.scss` to follow the PF4 technique of using CSS
variables in places where values need to be kept in sync.  In this case,
the height of the masthead (at 60px) is a CSS variable and, in future,
be changed in one place.

Change-Id: I648875672c6207da8f64af2985c6f9d51918e873
Bug-Url: https://bugzilla.redhat.com/1769306
Signed-off-by: Scott J Dickerson <sdickers@redhat.com>