runner: Synchronize the execute action
There is race condition in Runner-service playbook start call. At the
begging of the start call, the runner clean the artifacts dir with
cached data like setting, cmdline and extravars. So in case there two
parallel execution of the playbook, it may happen that this code is
executed for both playbooks, and the second playbook will use cached
data of the first one. This patch add synchronization block to start
playbook call so it can't happen.

Change-Id: Ia31f57bd7d179e10c160f8205f3269ee9bacfcac
Bug-Url: https://bugzilla.redhat.com/1787195
Signed-off-by: Ondra Machacek <omachace@redhat.com>