core: Fixed Import VM validation for Direct LUN
When importing a VM with the virtio-scsi driver disabled
from a Storage Domain after detaching and reattaching the
Storage Domain, the validation functionality would fail
to find the virtio-scsi driver and fail the import.

This issue was fixed by adding checking to ensure the
Device Interface is a virtio-scsi driver interface before
validating that the driver exists. In addition, in order to
fix this issue, a further modification that sets the chipset
of the bios of the VM from the cluster was added in order to
address a regress.

Change-Id: I6a601160dd5321997f315310ed6cafe7f9ddfcbf
Bug-Url: https://bugzilla.redhat.com/1788424
Signed-off-by: Steven Rosenberg <srosenbe@redhat.com>