webadmin: fix auto-attaching cd in change cd
When we auto-attach the windows guest tools CD
to a windows VM, the cd-name has prefix for location.
Therefore we couldn't see that the CD was attached in
Change CD dialog.

This change will let the string of the CD attached to
VMs to be equal or with the exact same suffix.

Change-Id: I1b8abf8b65c77b9a29cbd43b3541983dd30af28a
Bug-Url: https://bugzilla.redhat.com/1768393
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>