engine: Numa config CPUs needs to include all CPUs
The domainXML sent to libvirt when the VM is launched
was sending gaps when one hotplugged the total number
of CPUs while the number of NUMA CPUs in the list were
not being increased automatically. This caused a few
sets of failures.

Now the NUMA Configuration section shall include the range of
all Virtual CPUs from 0 to the maximum VCPU as sent in the
VCPU section based not just upon the Total Number in CPUs
variable, but when descrepant, the number of actual CPUs in
the Numa List.

Change-Id: Icf1763f7e5bb1308377e50173e82398f0beba6ac
Bug-Url: https://bugzilla.redhat.com/1437559
Signed-off-by: Steven Rosenberg <srosenbe@redhat.com>