core: remove agent version for WGT
On old windows VMs the Agent version was
in correlation with the Tools version.
The tools is the one to determaine the correct
version of WGT.

This patch will remove the consideration to
the reported Agent version.

As a result, the WGT ISO won't be attached
when the VM WGT is updated.

Change-Id: I49612595d3328ce03ae4f44ee569216d50682758
Bug-Url: https://bugzilla.redhat.com/1795246
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>