core: prevent setting RAW disks as supporting incremental backup
Currently, both RAW and COW disks can be set as 'incremental backup
enabled'.
This is wrong since incremental backup supported only for qcow2 disks.

This patch allows the user to set the disk as backup=incremental for COW
disks only when editing the disks.

For creating new disk via the UI - if the disk is set as 'enable
incremental backup' then the format that will be selected for the disk
is COW.
For editing existing disks via the UI, the checkbox for setting disk as
'enable incremental backup' is now grayed-out for RAW disks.

screenshot - https://imgur.com/a/mExzogT

For creating new disk via REST-API the user should specify the disk
configuration. In case of creating RAW disk with backup=incremental a
validation error will be raised, the same is true also for editing existing
disks via REST-API.

Change-Id: Ic3228373a00e8c61a02075dae53b2bfd5ac35d5b
Bug-Url: https://bugzilla.redhat.com/1793481
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>