Core: Hotplugging should distribute CPUs evenly
When updating the total number of CPUs during Hotplugging,
the additional CPUs would all be added to one Numa Node
due to the offline distribution of adding the offline cpus
sequentially.

Fixing the distribution required sorting the numa nodes
by the size of the cpu list to ensure the cpu additions are
first placed on the smaller list first and the cpu indexes
are distributed across each numa node evenly so that
adding additional CPUs provides for an even distribution.

Change-Id: I3f11c5cfc58cf74fcbcbf93fbb645180eee8aa09
Bug-Url: https://bugzilla.redhat.com/1177261
Signed-off-by: Steven Rosenberg <srosenbe@redhat.com>