engine: Disallow VM names ending with period
Libvirt does not support VM Names that end with a period
even though the ovirt engine allows the user to create or
modify a VM name with a period at the end. Therefore, when
attempting to run such a VM, the load fails.

Modified the regular expression validation to disallow adding
or modifying a VM name that ends with a period, thus avoiding
confusion for the user.

Change-Id: If583ece43569b57f2f5893325eb20f7dab79bd2b
Bug-Url: https://bugzilla.redhat.com/1721804
Signed-off-by: Steven Rosenberg <srosenbe@redhat.com>