engine: Set blank template for cloned VM edit
The clone of a VM created from a template failed during UpdateVmCommand.
This was caused by setting the VM's template as Blank but not
propagating this change to the parameters for UpdateVmCommand. The
command than failed with the following message:

The requested template doesn't belong to the same base template
as the original template.

Change-Id: Ie560fed4682de39373f9096e3fbb1f471b4009b8
Bug-Url: https://bugzilla.redhat.com/1800947
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>