kubevirt: Change Provider Commands to be non-transactive
Kubevirt flows assumes the provider exists in the database once it was
added in the UI dialog. However, the Add/Update ProviderCommand will
commit the transaction after its execution, causing sequential actions
in different transaction not be able to access the provider.

By modifying the command to non-transactive, the provider will be
accessible for other commands / actions.

Change-Id: I31cd64e22250e625a3c9b78f51432c46140cf462
Signed-off-by: Moti Asayag <masayag@redhat.com>