spec: Move selinux from engine to setup-base
Move /usr/share/ovirt-engine/selinux from the package ovirt-engine to
the package ovirt-engine-setup-base. This directory is used by
engine-setup's selinux plugin, which is packaged in the latter package.
Without this patch, installing only setup-base and running engine-setup
fails. This is needed e.g. for setup of the engine and dwh on two
separate machines.

Change-Id: I5298f438fcacbd7add620205013cdd639803a617
Bug-Url: https://bugzilla.redhat.com/1801129
Signed-off-by: Yedidyah Bar David <didi@redhat.com>