core: prevent RAW incremental backup supported disks on snapshot removal
RAW disks cannot be set as Backup=DiskBackup.Incremental, therefore, for
existing RAW disks, the user must create a snapshot for those disks that
will create a qcow2 layer on top of the RAW disks in order to mark them as
Backup=DiskBackup.Incremental.

But in case all the snapshots for the disk will be removed, the format
of it is again RAW, so the Backup=DiskBackup.Incremental must be
changed also to Backup=DiskBackup.None to enforce the rule for not
allowing incremental backup for RAW disks.

Change-Id: I7b18263b3fc44a7d33ced17fd5e94bb14b7d3547
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>