frontend: introduce attach WGT checkbox to runonce
This patch will introduce to the run once the option
of a windows VMs.
This option will let the user decide wether he wish
to have a secondary disk with windows guest tools.

This is useful case since the floppy drive is removed
and the user is installing windows VM. The drivers
will be available in setup time without changing disks.

Change-Id: Iff66dd8f13ea92c11abe9a31f9c96b7b4a10cf9f
Bug-Url: https://bugzilla.redhat.com/1798403
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>