webadmin: add a warning when removing a snapshot for losing backup support
Disk support for the incremental backup will be removed
in case of a snapshot removal for RAW based disks.

Implemented in the following patch -
https://gerrit.ovirt.org/#/c/106875/.

This patch adds a warning in the UI for the user to be aware of that
situation before confirming the snapshot removal.

Screenshot - https://imgur.com/a/zHQsstk

Change-Id: Iaaeb9980e16115396b6bc8c08747f5c0b20e397a
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>