core: introduce WGT secondary disk
This patch will introduce to the run once the option
of a windows VMs.
This option will let the user decide wether he wish
to have a secondary disk with windows guest tools.
The new CD-ROM is added for run-once and won't be saved
to the DB, therefore it will be auto-deleted from the VM.

This is useful case since the floppy drive is removed
and the user is installing windows VM. The drivers
will be available in setup time without changing disks.

Change-Id: I2b3d736384cabb2ed8ff88ddb45bf51afa622f3a
Bug-Url: https://bugzilla.redhat.com/1798403
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>