engine: show template sealed status in UI
When creating a template, it is possible to select an option to seal it.
However, the selected option is not stored in the DB and user does not
know which template has been sealed and which not.

This patch adds a new column to vm_static table is_template_sealed. The seal
status is then displayed on the main templates table. The seal status
could be changed in the template edit dialog.

This patch allows also to filter the templates by the seal status and
adds the field to the template when exporting / importing.

Change-Id: Id632b319cbab8f2f23ef532a52667ca590be74d5
Bug-Url: https://bugzilla.redhat.com/1512838
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>