WebAdmin: Added Bios Type to model initialization
When creating a new VM from a template where the Bios
Type is say set to any value other than the default,
the default Bios Type is used instead of the Bios Type
of the template. This is because the Bios Type was not
being set to the model from the template as per other
fields.

Now, when loading the model values from the template,
we include the setting of the Bios Type when initializing
the new VM values so that the new VM shall inherit the
correct Bios Type from the template it is based upon.

Change-Id: Idbab92dcaa206f4de6ef9c5980948e2ad8a613cb
Bug-Url: https://bugzilla.redhat.com/1678262
Signed-off-by: Steven Rosenberg <srosenbe@redhat.com>