webadmin: avoid importing a VM with an existing name
When importing a VM from a data domain and in case there's already VM
exists with that name, the operation starts and fails due to a
"VM Name already exists in the Data Center" error.
This patch blocks the operation on the UI level and continues once
the name entered is a unique one.

Screenshot: https://imgur.com/a/oaO6X14

Change-Id: I5984fce02ea85145997e419d9cdf5d5da77036f9
Bug-Url: https://bugzilla.redhat.com/1431535
Signed-off-by: Shani Leviim <sleviim@redhat.com>