core: Error due to usb correlation failure
An error occured due to a failure to correlate the existing
usb controller with usb controller of the domain xml returned
from libvirt. This was due to the engine preventing the alias
from containing the GUID of the existing usb device during
the launching of the VM. This resulted in the error and creating
a duplicate usb controller device for the VM. Warnings also occured
due to aliases that did not contain GUIDs and were not controls
the engine can correlate.

Removed the code preventing the GUID from being set as an alias
when launching the VM so that the alias of the usb controller can
contain the GUID. This way when libvirt resends the domain xml
when responding to a dumpxml request, the alias of the usb will
contain the GUID so that the monitor of the engine will find the
correct usb controller device without error and without creating a
duplicate device. Also removed warnings for channel and controller
devices that the engine cannot correlate to by filtering these
devices.

Change-Id: Id1d42eb18bef2cf9ed4f60ab361efed719b4dfb7
Bug-Url: https://bugzilla.redhat.com/1775248
Signed-off-by: Steven Rosenberg <srosenbe@redhat.com>