engine: match erratas for vm by fqdn and id
This patch extends current capabilities to find content hosts registered
in Red Hat Satellite / Foreman by using more search criteria.
1) host/vm id (facts.dmi::system::uuid))
2) vm fqdn (facts.network::fqdn))
3) host/vm name (facts.network::hostname)), hostname fact left for
   backward compatibility

The order of the checks is as stated above. The algorithm checks if
given criteria value is provided (not null) and executes query against
Satellite/Foreman. If content host is not resolved with given criteria,
the next in chain is used. Otherwise content host is returned.

The motivation behind using multi step checks is that potential changes
in vm / host like  a change in fqdn is reflected in Satellite/Foreman with a
delay. The only non-changing property is ID.

For matching erratas for engine host only engine host is used (as
before))

Change-Id: I3cc9e743fa8d7c5dcf859507cad38ab351406870
Bug-Url: https://bugzilla.redhat.com/1663217
Signed-off-by: Artur Socha <asocha@redhat.com>