core: Ensure at least one numa node is sent
One numa node is required to be sent within the
domain xml when launching a VM with memory hotplugging
when the user does not add one or more numa nodes.
In scenarios where no numa nodes were being sent,
but the max memory is greater than the current memory
and there are no offline CPUs, launching the VM would
fail.

Modifications now ensure that if the user did not
specify at least one numa node and if there is memory
hotplugging, that a default numa node will always be sent.

Change-Id: I21f2914a9538421a2646466be73d338899c719f6
Bug-Url: https://bugzilla.redhat.com/1808788
Signed-off-by: Steven Rosenberg <srosenbe@redhat.com>