webadmin: fix fetch 'in-preview' snapshot commit
If a snapshot is in a preview state, and the
page refreshes, committing the snapshot will
raise an error, because getSelectedItem()
returns null.

Each page state change (refresh for example)
will not affect the selected IN_PREVIEW
snapshot.

This patch fixes the correct selected
snapshot name for the description in the
popup message.

Older patch:
https://gerrit.ovirt.org/#/c/104436/

Screenshot: https://imgur.com/a/ZPZBWiP

Change-Id: Ic609ddfc8b7e214e02ae978d6e9ca6ed876bab59
Bug-Url: https://bugzilla.redhat.com/1743690
Signed-off-by: Ahmad Khiet <akhiet@redhat.com>