core: save checkpoint XML after starting a VM backup with checkpoint
When a VM backup is started a new checkpoint is created by Libvirt.
The engine needs to persist the checkpoint XML in his database for
a later use by Libvirt.

This patch adds a new column in vm_checkpoints table and saves the
checkpoint XML for each VM checkpoint after the backup was successfully
started.

Change-Id: I0975a06481677ce8211530278eb4ea4f72952083
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>