webadmin: Updated emulated machine in Run once
If the custom emulated machine is not set for the VM, the Run once
dialog automatically sets its value to cluster emulated machine value.
However, it causes the Vm to fail because the effective emulated machine is
determined in the RunVmCommand only if the custom emulated machine is null.

This patch removes the setting of the cluster value in the RunOnceModel,
instead it leaves it to the backend to determine the right effective
value.

Change-Id: I9851f872d32e4f4582465a6d0880cc16226c6cd2
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>