core: Mixing Q35 with Legacy Bios fails VM launch
When the Custom Emulation Machine Type is Q35 and the Bios Type
is Legacy Bios Type, the VM fails to launch after the initial
launch. This was due to the that Default Bios Type was
not being set when the Host is first brought up.

Fixed the Bios Type Default when a Host comes Up by updating
the Cluster's Bios Type to Q35 Legacy Bios when the Cluster's
Architection is updated to x86 to ensure the default
Bios Type will correlate with the default Emulation Machine
Type for x86 Architecture.

Change-Id: I9f014f25995b3247f06fe08050f52f6dc63dbb2d
Bug-Url: https://bugzilla.redhat.com/1813694
Signed-off-by: Steven Rosenberg <srosenbe@redhat.com>