core: Disable running VM with empty block hostdev path
If the path to block device for a hostdev SCSI disk is null, a device cannot
be used with a SCSI driver other that scsi_generic. Such VM XML won't be
accepted by libvirt.

If the VM has:
1. SCSI host device(s) added.
2. The blockPath in the corresponding HostDevice is null.
3. The custom property scsi_hostdev has a value different than scsi_generic.

Then an error is shown and the VM is not able to be run.

Change-Id: I8894539107678097b42139c1209b1e33382b4e87
Bug-Url: https://bugzilla.redhat.com/1801206
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>