core: escape new lines within OVF
There are values that new lines are critical
for them, such as custom script for cloud init.
The yaml we write into OVF will lose the new lines
and become incorrect on import.

This patch will escape the new line to OVF
and do the opposite when reading it.

For OVA, the format did remove all new lines
to fit xmllint, but it doesn't seem harmful
to leave them.

Change-Id: If7807cd618e24e10971a11f11dbd23ffb449c3af
Bug-Url: https://bugzilla.redhat.com/1755518
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>