core: re-arrange StartVmBackup to facilitate failure handling
This patch facilitates the structure of StartVmBackupCommand that uses
CoCo framework, it is needed to reduce code complexity and to detect
failures and handle them correctly.

Before that change, each phase automatically updated the command phase
to be the next one before the operation of the current phase
took place, so in case of failure in the middle of one of the phases
the command will jump to the next phase and will not return to the
interrupted phase in order to complete it.

Change-Id: I685a3dcc96fcc0aa470210fe3ca66f02e2de7ee6
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>