core: add recovery mechanism for engine crash during VM backup
In case that the engine crashes during VM backup operation after
'start_backup' operation already called on the host, when the engine
will get back up, it will try to re-run the 'start_backup' operation
and will get an error because the backup already started.

In this specific case, the engine can test if the backup already
initiated on the host and fetch the needed data so it will be able to
continue in the regular flow.

This patch adds a recovery mechanism for this case.

Change-Id: I736536c4fe3f6d90050f26b4ec28d6926c417c2a
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>