core: remove host - handle runner_on_unreachable event
Handle the case when due to the host being unreachable, ansible is not
able to add it to its inventory and the playbook is not started.

In this case, engine should mark the host as unreachable as well because
this is how it is marked by ansible in the playbook log.

See also: I7a0899d04e9b19f4c5729aa7c80844f96ae76c1a

Change-Id: I8aec00c9985343ceeb77316527e9005010985ec3
Signed-off-by: Eitan Raviv <eraviv@redhat.com>