core, webadmin: set storagePoolId on template registration
When registering a template from a data storage domain,
using a template's name which already exists, the operation should fail.

Currently, the storage pool id on getParameters() was set to zeros
instead of the right value, so the current validation is wrong.
This patch sets the storage pool id by the destination target.

screenshot: https://imgur.com/a/pRuix5X

Change-Id: Id62a8e6faacadb12d4ae72506a8be54e6f3c4ed2
Bug-Url: https://bugzilla.redhat.com/1816519
Signed-off-by: Shani Leviim <sleviim@redhat.com>