webadmin: Moved VmType copying to a builder
The Clone VM action did not copy VM type property explicitly - that
caused that the VM type was not set to the Edit dialog properly (rather,
default was selected).

This patch adds the VM type property to CoreVmBaseToUnitBuilder so that
it is copied with other properties. Note that CoreUnitToVmBaseBuilder
already contained the Vm type property.

Change-Id: I0f961f83cb8ddff7dc97eb974a4864456ebd8ff4
Bug-Url: https://bugzilla.redhat.com/1817085
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>