core: add multipath status in engine log
Multipath status logs were logged only when
the host re-activated, MultipathHealthHandler
was in GetStatsAsyncVDSCommand.java only,
GetStatsAsyncVDSCommand.java is called when
the host re-activated.

The only logged message was that the
domain is in problem 'PROBLEMATIC' from
the IrsProxy.

Multipath status changes now appear in the
engine logs and in the Events tab, by
adding it to GetStatsAsyncVDSCommand.java
which is called periodically.

Screenshots: https://imgur.com/a/TI7pmZj

Change-Id: Iec0d53d07f4c556b5c71d29d0826c19416da8d43
Bug-Url: https://bugzilla.redhat.com/1786999
Signed-off-by: Ahmad Khiet <akhiet@redhat.com>