core: add recover mechanism for missing checkpoint XML
When starting a backup operation that includes a checkpoint creation,
VDSM should return the checkpoint XML info that returns from libvirt to
the Engine. The checkpoint XML is stored in the Engine database for
later use.

The checkpoint XML info can be missing if:
  1. The backup contains only RAW disks so checkpoint isn't needed.
  2. An error occurred while the host tries to fetch the checkpoint XML.

In case an error has occurred during 'start_backup' operation in the host
and the checkpoint XML is missing, the Engine will try to fetch the
checkpoint XML again by calling 'backup_info'. If the checkpoint XML is
still missing, the Engine will fail the backup operation and the chain
cannot be used anymore so a full backup should be taken.

Change-Id: Ia068a5fb988803c1b73fe80b83efd9a00c18252b
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>