pythonlib: Fix str x bytes conflicts in TicketEncoder.encode
TicketEncoder.encode currently can't work since it mixes strings and
bytes freely and passes incorrect types to the underlying libraries.
Let's unify the types to make the method working.  The method argument
type and the return type are unchanged and the apparently only caller
of the method, ovirt-vmconsole-list.py, is happy with the types.

Change-Id: I53dff7a757e791a95867e8390039b464ee016e82
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>