ansible: don't start imageio service
imageio is stared by systemd automatically, as vdsm wants imageio
service, so there's no need to start it explicitely from ansible
role. Remove image from list of services to be started.

Also, name of the service is ovirt-imageio. Fix the name of the
service in list of service to be stopped.

Change-Id: Id1f1c7de63084f5af57045012ab72f552464a97c
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>