engine: Make NUMA work with hugepages
The engine did not support numa with hugepages - the free hugepages
were not stored in the numa node statistics and the scheduler
did not consider hugepages when calculating available memory.

This patch stores free hugepages reported by vdsm per numa node to
numa_node table. The NumaPinningHelper reads the statistics and
considers it when calculating available memory for the virtual machine.

Change-Id: Ib9e7253120aeead0d0fbdbb271610d35b190d358
Bug-Url: https://bugzilla.redhat.com/1812316
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>