core: Fixed Bios type defaults
When the Custom Emulation Machine Type is Q35 and the Bios Type
is Legacy Bios Type, the VM fails to launch after the initial
launch. This was due to the fact that the Default Bios Type was
not being reset when the Host is first brought up.

Fixed the Bios Type Default when a Host comes Up by updating
the Cluster's Bios Type to Q35 Legacy Bios when the Cluster's
Architection is updated to x86 to ensure the default
Bios Type will correlate with the default Emulation Machine
Type for x86 Architecture.

This fix also ensures the the Bios Type will be set properly
when the architecture is not set to undefined and the Bios Type
is set to its default.

Change-Id: Ibf1e87a90876bfa861766452292c208c3ed4297e
Bug-Url: https://bugzilla.redhat.com/1813694
Signed-off-by: Steven Rosenberg <srosenbe@redhat.com>